do not show buttons for forbidden actions

+++ src/components/Comment/CommentControls.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+    import { getContext } from 'svelte';
     import { _ } from 'svelte-i18n';
 	import { createEventDispatcher } from 'svelte';
 
@@ -10,6 +11,8 @@
     export let status: CommentStatus;
 
 	const dispatch = createEventDispatcher();
+
+    const { allowEdit, allowDelete } = getContext("commenti-settings");
 </script>
 
 
@@ -22,17 +25,21 @@
         </button>
 
         {#if $authState.user.username == valueIfExists(status, comment.author?.username)}
-            <button
-                on:click={() => dispatch('delete')}
-                class="commenti-inline-button">
-                {$_("commentControls.delete")}
-            </button>
+            {#if $allowDelete}
+                <button
+                    on:click={() => dispatch('delete')}
+                    class="commenti-inline-button">
+                    {$_("commentControls.delete")}
+                </button>
+            {/if}
 
-            <button
-                on:click={() => dispatch('edit')}
-                class="commenti-inline-button">
-                {$_("commentControls.edit")}
-            </button>
+            {#if $allowEdit}
+                <button
+                    on:click={() => dispatch('edit')}
+                    class="commenti-inline-button">
+                    {$_("commentControls.edit")}
+                </button>
+            {/if}
         {/if}
     {/if}
 </div>
